Moved JdbcTemplateMapper package
